+++ packages/ethereum-indexer-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # ethereum-indexer-server
 
+## 0.6.28
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer@0.6.19
+  - ethereum-indexer-db-utils@0.6.19
+  - ethereum-indexer-fs-cache@0.6.19
+  - ethereum-indexer-utils@0.6.11
+
 ## 0.6.27
 
 ### Patch Changes